+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.0.1] - 2019-01-02
 ### Added
 - Added changelog.
 - Added core library.
+- Added API Helper functions
+
+### Changed
+- Format of model validation errors